Added parent page title to all mobile navigation
The way the site was set up is a little inconsistent. Different pages
have their titles inserted in different ways with different classes.

This is an interim solution so that all pages have the parent page
title inserted at the top tot mobile sub navigation menu drop downs.

I suggest a few days to make the templates consistent and to remove
hard-coded titles so that when changes are made from within the admin
that they are reflected in page titles dynamically.